CDN SDK Auto Released By wb-czb515173,Version：1.34.51

+++ cdn/include/alibabacloud/cdn/model/DescribeTopDomainsByFlowResult.h
@@ -34,7 +34,7 @@ namespace AlibabaCloud
 			public:
 				struct TopDomain
 				{
-					long maxBps;
+					float maxBps;
 					std::string domainName;
 					std::string maxBpsTime;
 					long rank;
